tests: moved testStorageQuota fn
